+++ app/api/delete/temp-storage/text/[id]/[uid]/route.tsx
@@ -3,7 +3,7 @@ import { createClient } from "@/lib/supabase/server";
 
 export async function DELETE(
   req: NextRequest,
-  { params }: { params: { id: string, filename: string} }
+  { params }: { params: { id: string} }
 ) {
   const supabase = await createClient();
 
@@ -17,22 +17,20 @@ export async function DELETE(
   }
 
   // Delete record from temp_storage and file from storage
-  const { id, filename } = await params;
-  const decodedName = decodeURIComponent(filename);
+  const { id } = await params;
   
-  const response = await supabase
+  const { data: deleteData, error: deleteError } = await supabase
   .from('temp_storage')
   .delete()
   .eq('id', id)
+  .eq('uid', user.id);
 
-  const { data, error } = await supabase
-  .storage
-  .from('temporary_storage')
-  .remove([decodedName])
-  
-  if (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 });
-  } 
+  if (deleteError) {
+    return NextResponse.json(
+      { error: deleteError.message },
+      { status: 500 }
+    );
+  }
 
-  return NextResponse.json({ message: "Record deleted", response });
+  return NextResponse.json({ message: "Record deleted" });
 }
